@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
 
 namespace UtilityExtensions.Core.Configurations.Attributes
 {
@@ -14,13 +10,21 @@ namespace UtilityExtensions.Core.Configurations.Attributes
         public BooleanConfigProperty(string @default, string newName = null, string trueValue = "1") : base(@default, newName)
         {
             if (string.IsNullOrEmpty(trueValue))
+            {
                 throw new ArgumentNullException(nameof(trueValue));
+            }
 
             this.trueValue = trueValue;
         }
 
-        public override object ConvertFromString(string s) => (s?.ToLower() == trueValue?.ToLower() || s?.ToLower() == "true");
+        public override object ConvertFromString(string s)
+        {
+            return (s?.ToLower() == trueValue?.ToLower() || s?.ToLower() == "true");
+        }
 
-        public override string ConvertToString(object o) => o is bool value ? value ? "1" : "0" : @default;
+        public override string ConvertToString(object o)
+        {
+            return o is bool value ? value ? "1" : "0" : @default;
+        }
     }
 }
