@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace UtilityExtensions.Core.Configurations.Attributes
 {
@@ -14,8 +10,14 @@ namespace UtilityExtensions.Core.Configurations.Attributes
         {
         }
 
-        public override object ConvertFromString(string s) => s ?? @default;
+        public override object ConvertFromString(string s)
+        {
+            return s ?? @default;
+        }
 
-        public override string ConvertToString(object o) => o?.ToString() ?? @default;
+        public override string ConvertToString(object o)
+        {
+            return o?.ToString() ?? @default;
+        }
     }
 }
