@@ -21,15 +21,15 @@ namespace UtilityExtensions.Core.Geometry
             get => ClampAngle(value); private set => this.value = ClampAngle(value);
         }
 
-        public double Rad
-        {
-            get => value * (Math.PI / 180);
-        }
+        public double Rad => value * (Math.PI / 180);
 
         public bool InsideRange(Angle start, Angle end)
         {
             if (start.Value < end.Value)
+            {
                 return start.Value <= Value && Value <= end.Value;
+            }
+
             return start.Value <= Value || Value <= end.Value;
         }
 
@@ -53,7 +53,9 @@ namespace UtilityExtensions.Core.Geometry
             angle = angle % 360;
 
             if (angle < 0)
+            {
                 angle += 360;
+            }
 
             return angle;
         }
