@@ -4,6 +4,9 @@ namespace UtilityExtensions.Core.Geometry.Collision
 {
     public class RectangleCollidesRectangle : IShapesCollision<Rectangle, Rectangle>
     {
-        public bool CheckCollision(Rectangle rect1, Rectangle rect2) => GeometryExtensions.IsShapes2DIntersecting(rect1, rect2);
+        public bool CheckCollision(Rectangle rect1, Rectangle rect2)
+        {
+            return GeometryExtensions.IsShapes2DIntersecting(rect1, rect2);
+        }
     }
 }
