@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using UtilityExtensions.Extensions;
 
 namespace UtilityExtensions.Core
@@ -25,73 +23,175 @@ namespace UtilityExtensions.Core
         public float Float => (float)value;
         public decimal Decimal => (decimal)value;
 
-        public static implicit operator long(Numeric n) => n.Long;
+        public static implicit operator long(Numeric n)
+        {
+            return n.Long;
+        }
 
-        public static implicit operator Numeric(long n) => new Numeric(n);
+        public static implicit operator Numeric(long n)
+        {
+            return new Numeric(n);
+        }
 
-        public static implicit operator int(Numeric n) => n.Int;
+        public static implicit operator int(Numeric n)
+        {
+            return n.Int;
+        }
 
-        public static implicit operator Numeric(int n) => new Numeric(n);
+        public static implicit operator Numeric(int n)
+        {
+            return new Numeric(n);
+        }
 
-        public static implicit operator double(Numeric n) => n.Double;
+        public static implicit operator double(Numeric n)
+        {
+            return n.Double;
+        }
 
-        public static implicit operator Numeric(double n) => new Numeric(n);
+        public static implicit operator Numeric(double n)
+        {
+            return new Numeric(n);
+        }
 
-        public static implicit operator float(Numeric n) => n.Float;
+        public static implicit operator float(Numeric n)
+        {
+            return n.Float;
+        }
 
-        public static implicit operator Numeric(float n) => new Numeric(n);
+        public static implicit operator Numeric(float n)
+        {
+            return new Numeric(n);
+        }
 
-        public static implicit operator decimal(Numeric n) => n.Decimal;
+        public static implicit operator decimal(Numeric n)
+        {
+            return n.Decimal;
+        }
 
-        public static implicit operator Numeric(decimal n) => new Numeric((double)n);
+        public static implicit operator Numeric(decimal n)
+        {
+            return new Numeric((double)n);
+        }
 
-        public static Numeric operator +(Numeric n1, Numeric n2) => n1.value + n2.value;
+        public static Numeric operator +(Numeric n1, Numeric n2)
+        {
+            return n1.value + n2.value;
+        }
 
-        public static Numeric operator +(Numeric n1, long n2) => n1.value + n2;
+        public static Numeric operator +(Numeric n1, long n2)
+        {
+            return n1.value + n2;
+        }
 
-        public static Numeric operator +(Numeric n1, int n2) => n1.value + n2;
+        public static Numeric operator +(Numeric n1, int n2)
+        {
+            return n1.value + n2;
+        }
 
-        public static Numeric operator +(Numeric n1, double n2) => n1.value + n2;
+        public static Numeric operator +(Numeric n1, double n2)
+        {
+            return n1.value + n2;
+        }
 
-        public static Numeric operator +(Numeric n1, float n2) => n1.value + n2;
+        public static Numeric operator +(Numeric n1, float n2)
+        {
+            return n1.value + n2;
+        }
 
-        public static Numeric operator +(Numeric n1, decimal n2) => n1.value + (double)n2;
+        public static Numeric operator +(Numeric n1, decimal n2)
+        {
+            return n1.value + (double)n2;
+        }
 
-        public static Numeric operator -(Numeric n1, Numeric n2) => n1.value - n2.value;
+        public static Numeric operator -(Numeric n1, Numeric n2)
+        {
+            return n1.value - n2.value;
+        }
 
-        public static Numeric operator -(Numeric n1, long n2) => n1.value - n2;
+        public static Numeric operator -(Numeric n1, long n2)
+        {
+            return n1.value - n2;
+        }
 
-        public static Numeric operator -(Numeric n1, int n2) => n1.value - n2;
+        public static Numeric operator -(Numeric n1, int n2)
+        {
+            return n1.value - n2;
+        }
 
-        public static Numeric operator -(Numeric n1, double n2) => n1.value - n2;
+        public static Numeric operator -(Numeric n1, double n2)
+        {
+            return n1.value - n2;
+        }
 
-        public static Numeric operator -(Numeric n1, float n2) => n1.value - n2;
+        public static Numeric operator -(Numeric n1, float n2)
+        {
+            return n1.value - n2;
+        }
 
-        public static Numeric operator -(Numeric n1, decimal n2) => n1.value - (double)n2;
+        public static Numeric operator -(Numeric n1, decimal n2)
+        {
+            return n1.value - (double)n2;
+        }
 
-        public static Numeric operator *(Numeric n1, Numeric n2) => n1.value * n2.value;
+        public static Numeric operator *(Numeric n1, Numeric n2)
+        {
+            return n1.value * n2.value;
+        }
 
-        public static Numeric operator *(Numeric n1, long n2) => n1.value * n2;
+        public static Numeric operator *(Numeric n1, long n2)
+        {
+            return n1.value * n2;
+        }
 
-        public static Numeric operator *(Numeric n1, int n2) => n1.value * n2;
+        public static Numeric operator *(Numeric n1, int n2)
+        {
+            return n1.value * n2;
+        }
 
-        public static Numeric operator *(Numeric n1, double n2) => n1.value * n2;
+        public static Numeric operator *(Numeric n1, double n2)
+        {
+            return n1.value * n2;
+        }
 
-        public static Numeric operator *(Numeric n1, float n2) => n1.value * n2;
+        public static Numeric operator *(Numeric n1, float n2)
+        {
+            return n1.value * n2;
+        }
 
-        public static Numeric operator *(Numeric n1, decimal n2) => n1.value * (double)n2;
+        public static Numeric operator *(Numeric n1, decimal n2)
+        {
+            return n1.value * (double)n2;
+        }
 
-        public static Numeric operator /(Numeric n1, Numeric n2) => n1.value / n2.value;
+        public static Numeric operator /(Numeric n1, Numeric n2)
+        {
+            return n1.value / n2.value;
+        }
 
-        public static Numeric operator /(Numeric n1, long n2) => n1.value / n2;
+        public static Numeric operator /(Numeric n1, long n2)
+        {
+            return n1.value / n2;
+        }
 
-        public static Numeric operator /(Numeric n1, int n2) => n1.value / n2;
+        public static Numeric operator /(Numeric n1, int n2)
+        {
+            return n1.value / n2;
+        }
 
-        public static Numeric operator /(Numeric n1, double n2) => n1.value / n2;
+        public static Numeric operator /(Numeric n1, double n2)
+        {
+            return n1.value / n2;
+        }
 
-        public static Numeric operator /(Numeric n1, float n2) => n1.value / n2;
+        public static Numeric operator /(Numeric n1, float n2)
+        {
+            return n1.value / n2;
+        }
 
-        public static Numeric operator /(Numeric n1, decimal n2) => n1.value / (double)n2;
+        public static Numeric operator /(Numeric n1, decimal n2)
+        {
+            return n1.value / (double)n2;
+        }
 
         public static bool operator ==(Numeric left, Numeric right)
         {
