@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace UtilityExtensions.Core.Services
+﻿namespace UtilityExtensions.Core.Services
 {
     public interface IService
     {
