@@ -1,7 +1,6 @@
 ﻿using NUnit.Framework;
 using System.Collections.Generic;
 using UtilityExtensions.Core;
-using UtilityExtensions.Extensions;
 
 namespace UtilityExtensions_Test
 {
@@ -9,22 +8,22 @@ namespace UtilityExtensions_Test
     {
         private class MapConvertionTest
         {
-            public string stringTest { get; set; } = "test";
-            public int intTest { get; set; } = 42;
-            public decimal decimalTest { get; set; } = 12.34m;
-            public NestedClass nestedClassTest { get; set; } = new NestedClass();
-            public List<decimal> listTest { get; set; } = new List<decimal> { 1.2m, 32 };
+            public string StringTest { get; set; } = "test";
+            public int IntTest { get; set; } = 42;
+            public decimal DecimalTest { get; set; } = 12.34m;
+            public NestedClass NestedClassTest { get; set; } = new NestedClass();
+            public List<decimal> ListTest { get; set; } = new List<decimal> { 1.2m, 32 };
 
             [MapIgnore]
-            public int ignore { get; set; }
+            public int Ignore { get; set; }
         }
 
         private class NestedClass
         {
-            public string nestedStringTest { get; set; } = "test2";
-            public int nestedIntTest { get; set; } = 21;
-            public decimal nestedDecimalTest { get; set; } = 100.1m;
-            public List<decimal> nestedListTest { get; set; } = new List<decimal> { 1.2m, 32 };
+            public string NestedStringTest { get; set; } = "test2";
+            public int NestedIntTest { get; set; } = 21;
+            public decimal NestedDecimalTest { get; set; } = 100.1m;
+            public List<decimal> NestedListTest { get; set; } = new List<decimal> { 1.2m, 32 };
         }
 
         [SetUp]
@@ -35,18 +34,18 @@ namespace UtilityExtensions_Test
         [Test, Order(0)]
         public void ConvertObjectToMap()
         {
-            var obj = new MapConvertionTest();
-            var convertMap = obj.ToMap();
-            var checkMap = new Map() {
-                { "stringTest" , "test" },
-                { "intTest" , 42 },
-                { "decimalTest" , 12.34m },
-                { "listTest" , new List<decimal> { 1.2m, 32 } },
-                { "nestedClassTest" ,  new Map(){
-                    {"nestedStringTest", "test2" },
-                    {"nestedIntTest", 21 },
-                    {"nestedDecimalTest", 100.1m },
-                    {"nestedListTest", new List<decimal> { 1.2m, 32 }},
+            MapConvertionTest obj = new MapConvertionTest();
+            Map convertMap = obj.ToMap();
+            Map checkMap = new Map() {
+                { "StringTest" , "test" },
+                { "IntTest" , 42 },
+                { "DecimalTest" , 12.34m },
+                { "ListTest" , new List<decimal> { 1.2m, 32 } },
+                { "NestedClassTest" ,  new Map(){
+                    {"NestedStringTest", "test2" },
+                    {"NestedIntTest", 21 },
+                    {"NestedDecimalTest", 100.1m },
+                    {"NestedListTest", new List<decimal> { 1.2m, 32 }},
                 } }
             };
 
@@ -56,32 +55,51 @@ namespace UtilityExtensions_Test
         [Test]
         public void ConvertMapToObject()
         {
-            var convertMap = new Map() {
-                { "stringTest" , null},
-                { "intTest" , 1 },
-                { "decimalTest" , 2m },
-                { "nestedClassTest" ,  new Map(){
-                    {"nestedStringTest", "test" },
-                    {"nestedIntTest", 3 },
-                    {"nestedDecimalTest", 4m },
+            Map convertMap = new Map() {
+                { "StringTest" , null},
+                { "IntTest" , 1 },
+                { "DecimalTest" , 2m },
+                { "NestedClassTest" ,  new Map(){
+                    {"NestedStringTest", "test" },
+                    {"NestedIntTest", 3 },
+                    {"NestedDecimalTest", 4m },
                 } }
             };
-            var obj = convertMap.FromMap<MapConvertionTest>();
+            MapConvertionTest obj = convertMap.FromMap<MapConvertionTest>();
             Assert.IsNotNull(obj);
             Assert.IsTrue(obj is MapConvertionTest);
-            Assert.IsNull(obj.stringTest);
-            Assert.IsTrue(obj.intTest == 1);
-            Assert.IsTrue(obj.decimalTest == 2);
-            Assert.IsNotNull(obj.nestedClassTest);
-            Assert.IsTrue(obj.nestedClassTest.nestedStringTest == "test");
-            Assert.IsTrue(obj.nestedClassTest.nestedIntTest == 3);
-            Assert.IsTrue(obj.nestedClassTest.nestedDecimalTest == 4);
+            Assert.IsNull(obj.StringTest);
+            Assert.IsTrue(obj.IntTest == 1);
+            Assert.IsTrue(obj.DecimalTest == 2);
+            Assert.IsNotNull(obj.NestedClassTest);
+            Assert.IsTrue(obj.NestedClassTest.NestedStringTest == "test");
+            Assert.IsTrue(obj.NestedClassTest.NestedIntTest == 3);
+            Assert.IsTrue(obj.NestedClassTest.NestedDecimalTest == 4);
+        }
+
+        [Test]
+        public void Clone()
+        {
+            Map checkMap = new Map() {
+                { "StringTest" , "test" },
+                { "IntTest" , 42 },
+                { "DecimalTest" , 12.34m },
+                { "ListTest" , new List<decimal> { 1.2m, 32 } },
+                { "NestedClassTest" ,  new Map(){
+                    {"NestedStringTest", "test2" },
+                    {"NestedIntTest", 21 },
+                    {"NestedDecimalTest", 100.1m },
+                    {"NestedListTest", new List<decimal> { 1.2m, 32 }},
+                } }
+            };
+            Map cloneMap = checkMap.Clone() as Map;
+            Assert.IsTrue(cloneMap == checkMap);
         }
 
         [Test]
         public void MapToString()
         {
-            var map = new Map()
+            Map map = new Map()
             {
                 {"field1", 10 },
                 {"field2", null },
@@ -101,12 +119,12 @@ namespace UtilityExtensions_Test
         [Test]
         public void CompareMaps_DifferentFieldValue_ReturnsFalse()
         {
-            var map1 = new Map()
+            Map map1 = new Map()
             {
                 {"field1", 10 },
                 {"field2", null },
             };
-            var map2 = new Map()
+            Map map2 = new Map()
             {
                 {"field1", 10 },
                 {"field2", "test" },
@@ -118,12 +136,12 @@ namespace UtilityExtensions_Test
         [Test]
         public void CompareMaps_SameFieldValues_ReturnsTrue()
         {
-            var map1 = new Map()
+            Map map1 = new Map()
             {
                 {"field1", 10 },
                 {"field2", "test" },
             };
-            var map2 = new Map()
+            Map map2 = new Map()
             {
                 {"field1", 10 },
                 {"field2", "test" },
@@ -135,12 +153,12 @@ namespace UtilityExtensions_Test
         [Test]
         public void CompareMaps_DifferentFields_ReturnsFalse()
         {
-            var map1 = new Map()
+            Map map1 = new Map()
             {
                 {"field1", 10 },
                 {"field2", "test" },
             };
-            var map2 = new Map()
+            Map map2 = new Map()
             {
                 {"field1", 10 },
                 {"field3", "test" },
@@ -152,11 +170,11 @@ namespace UtilityExtensions_Test
         [Test]
         public void CompareMaps_MissingFields_ReturnsFalse()
         {
-            var map1 = new Map()
+            Map map1 = new Map()
             {
                 {"field1", 10 },
             };
-            var map2 = new Map()
+            Map map2 = new Map()
             {
                 {"field1", 10 },
                 {"field2", "test" },
