@@ -1,6 +1,6 @@
 ﻿using NUnit.Framework;
 using System;
-using UtilityExtensions.Core;
+using System.Linq;
 using UtilityExtensions.Core.Validations;
 
 namespace UtilityExtensions_Test
@@ -12,9 +12,9 @@ namespace UtilityExtensions_Test
         {
             try
             {
-                var string1 = "123";
-                var string2 = "321";
-                var string3 = "";
+                string string1 = "123";
+                string string2 = "321";
+                string string3 = "";
 
                 ValidationManager<string>.Add(string1, nameof(string1))
                                          .Add(string2, nameof(string2))
@@ -34,9 +34,9 @@ namespace UtilityExtensions_Test
         {
             try
             {
-                var string1 = "123";
-                var string2 = "321";
-                var string3 = "12";
+                string string1 = "123";
+                string string2 = "321";
+                string string3 = "12";
 
                 ValidationManager<string>.UseSettings(new ValidationManager.Settings { throwExceptionOnFail = true, validateImmediately = true })
                                          .Add(string1, nameof(string1))
@@ -54,13 +54,39 @@ namespace UtilityExtensions_Test
         }
 
         [Test]
+        public void ValidateStrings_ValidEmail()
+        {
+            string string1 = "validemail@gmail.com";
+
+            var vm = ValidationManager<string>.UseSettings(new ValidationManager.Settings { throwExceptionOnFail = false, validateImmediately = true })
+                                     .Add(string1, nameof(string1))
+                                     .ValidEmail();
+
+            Assert.IsTrue(vm.FailedParameters().Count == 0);
+            Assert.IsTrue(vm.Parameters.First().Results.First().IsSuccess);
+        }
+
+        [Test]
+        public void ValidateStrings_InvalidEmail()
+        {
+            string string1 = "invalidmail.cm";
+
+            var vm = ValidationManager<string>.UseSettings(new ValidationManager.Settings { throwExceptionOnFail = false, validateImmediately = true })
+                                     .Add(string1, nameof(string1))
+                                     .ValidEmail();
+
+            Assert.IsTrue(vm.FailedParameters().Count == 1);
+            Assert.IsFalse(vm.Parameters.First().Results.First().IsSuccess);
+        }
+
+        [Test]
         public void ValidateDecimal_InRange()
         {
             try
             {
-                var decim1 = 30.23m;
-                var decim2 = -13m;
-                var decim3 = 2000m;
+                decimal decim1 = 30.23m;
+                decimal decim2 = -13m;
+                decimal decim3 = 2000m;
 
                 ValidationManager<decimal>.Add(decim1, nameof(decim1))
                                          .Add(decim2, nameof(decim2))
